extracted interfaces for AudioManager and VoiceManager and have them implement ConfigOptionProvider
