+++ bundles/core/org.eclipse.smarthome.core.audio/OSGI-INF/AudioManager.xml
@@ -9,12 +9,13 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" configuration-policy="optional" immediate="true" modified="modified" name="org.eclipse.smarthome.audio">
-   <implementation class="org.eclipse.smarthome.core.audio.AudioManager"/>
+   <implementation class="org.eclipse.smarthome.core.audio.internal.AudioManagerImpl"/>
    <reference bind="addAudioSource" cardinality="0..n" interface="org.eclipse.smarthome.core.audio.AudioSource" name="AudioSource" policy="dynamic" unbind="removeAudioSource"/>
    <reference bind="addAudioSink" cardinality="0..n" interface="org.eclipse.smarthome.core.audio.AudioSink" name="AudioSink" policy="dynamic" unbind="removeAudioSink"/>
 
    <service>
       <provide interface="org.eclipse.smarthome.core.audio.AudioManager"/>
+      <provide interface="org.eclipse.smarthome.config.core.ConfigOptionProvider"/>
    </service>
    <property name="service.pid" type="String" value="org.eclipse.smarthome.audio"/>
    <property name="service.config.description.uri" type="String" value="system:audio"/>
