+++ bundles/core/org.eclipse.smarthome.core.voice/OSGI-INF/VoiceManager.xml
@@ -9,7 +9,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" configuration-policy="optional" immediate="true" modified="modified" name="org.eclipse.smarthome.voice">
-   <implementation class="org.eclipse.smarthome.core.voice.VoiceManager"/>
+   <implementation class="org.eclipse.smarthome.core.voice.internal.VoiceManagerImpl"/>
    <reference bind="setAudioManager" cardinality="1..1" interface="org.eclipse.smarthome.core.audio.AudioManager" name="AudioManager" policy="static" unbind="unsetAudioManager"/>
    <reference bind="addKSService" cardinality="0..n" interface="org.eclipse.smarthome.core.voice.KSService" name="KSService" policy="dynamic" unbind="removeKSService"/>
    <reference bind="addSTTService" cardinality="0..n" interface="org.eclipse.smarthome.core.voice.STTService" name="STTService" policy="dynamic" unbind="removeSTTService"/>
@@ -19,6 +19,7 @@
 
    <service>
       <provide interface="org.eclipse.smarthome.core.voice.VoiceManager"/>
+      <provide interface="org.eclipse.smarthome.config.core.ConfigOptionProvider"/>
    </service>
    <property name="service.pid" type="String" value="org.eclipse.smarthome.voice"/>
    <property name="service.config.description.uri" type="String" value="system:voice"/>
